+++ packages/analyzer/src/model/warning.ts
@@ -90,8 +90,11 @@ export class Warning {
 
     let result = '';
     if (options.verbosity !== 'one-line') {
-      const underlined =
-          underlineCode(this.sourceRange, this._parsedDocument, colorize);
+      const underlined = underlineCode(
+          this.sourceRange,
+          this._parsedDocument,
+          colorize,
+          options.maxCodeLines);
       if (underlined) {
         result += underlined;
       }
@@ -180,6 +183,11 @@ export interface WarningStringifyOptions {
    * to the reported file.
    */
   readonly resolver?: UrlResolver;
+  /**
+   * If given, we will only print this many lines of code. Otherwise print all
+   * lines in the source range.
+   */
+  readonly maxCodeLines?: number;
 }
 const defaultPrinterOptions = {
   verbosity: 'full' as 'full',